Restructure the analyzer, put reference to it in test project
