+++ test/Microsoft.ML.StaticPipelineTesting/Microsoft.ML.StaticPipelineTesting.csproj
@@ -5,10 +5,6 @@
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\Microsoft.ML.Data\Microsoft.ML.Data.csproj" />
-
-    <ProjectReference Include="..\..\tools-local\Microsoft.ML.LibraryAnalyzer\Microsoft.ML.LibraryAnalyzer.csproj">
-      <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
-      <OutputItemType>Analyzer</OutputItemType>
-    </ProjectReference>
+    <ProjectReference Include="..\..\tools-local\Microsoft.ML.Analyzer\Microsoft.ML.Analyzer.csproj" />
   </ItemGroup>
 </Project>
