+++ tools-local/Microsoft.ML.Analyzer/Microsoft.ML.Analyzer.csproj
