+++ tools-local/Microsoft.ML.Analyzer/PigstyDelegateAnalyzer.cs
